restructuring the dashboard overview layout
