+++ src/web/components/Card/Card.tsx
@@ -1,0 +1,40 @@
+import React from 'react';
+import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
+import Paper from '@material-ui/core/Paper';
+import Grid from '@material-ui/core/Grid';
+
+const useStyles = makeStyles((theme: Theme) =>
+    createStyles({
+        root: {
+            flexGrow: 1
+        },
+        paper: {
+            padding: theme.spacing(2),
+            textAlign: 'center',
+            color: theme.palette.text.secondary
+        }
+    })
+);
+
+export default function AutoGrid() {
+    const classes = useStyles();
+
+    return (
+        <div className={classes.root}>
+            <Grid container spacing={3}>
+                <Grid item xs>
+                    <Paper className={classes.paper}>xs</Paper>
+                </Grid>
+                <Grid item xs>
+                    <Paper className={classes.paper}>xs</Paper>
+                </Grid>
+                <Grid item xs>
+                    <Paper className={classes.paper}>xs</Paper>
+                </Grid>
+                <Grid item xs>
+                    <Paper className={classes.paper}>xs</Paper>
+                </Grid>
+            </Grid>
+        </div>
+    );
+}
