+++ src/web/components/Main/Main.tsx
@@ -15,8 +15,6 @@ import ChevronRightIcon from '@material-ui/icons/ChevronRight';
 import ListItem from '@material-ui/core/ListItem';
 import ListItemIcon from '@material-ui/core/ListItemIcon';
 import ListItemText from '@material-ui/core/ListItemText';
-import InboxIcon from '@material-ui/icons/MoveToInbox';
-import MailIcon from '@material-ui/icons/Mail';
 import Badge from '@material-ui/core/Badge';
 import MenuItem from '@material-ui/core/MenuItem';
 import Menu from '@material-ui/core/Menu';
@@ -29,7 +27,7 @@ import Overview from '../../public/overview.svg';
 import Employees from '../../public/employees.svg';
 import Integrations from '../../public/integrations.svg';
 import DashboardLogo from '../../public/dashboardLogo.svg';
-import { Paper } from '@material-ui/core';
+import Card from '../Card/Card';
 
 const drawerWidth = 240;
 
@@ -85,7 +83,9 @@ const useStyles = makeStyles((theme: Theme) =>
             }
         },
         customizeToolbar: {
-            // minHeight: 100
+            minHeight: 100,
+            border: '0',
+            backgroundColor: '#F7F8FC'
         },
         title: {
             display: 'none',
@@ -155,10 +155,13 @@ const useStyles = makeStyles((theme: Theme) =>
             padding: theme.spacing(0, 1),
             // necessary for content to be below app bar
             ...theme.mixins.toolbar
+            // marginBottom: '25px'
         },
         content: {
             flexGrow: 1,
-            padding: theme.spacing(3)
+            padding: theme.spacing(3),
+            backgroundColor: '#F7F8FC',
+            marginTop: '70px'
         }
     })
 );
@@ -264,7 +267,12 @@ export default function Main() {
                     position="fixed"
                     className={clsx(classes.appBar, {
                         [classes.appBarShift]: open
-                    })}>
+                    })}
+                    style={{
+                        background: 'transparent',
+                        boxShadow: 'none',
+                        backgroundColor: '#F7F8FC'
+                    }}>
                     <Toolbar className={classes.customizeToolbar}>
                         <IconButton
                             color="inherit"
@@ -374,21 +382,9 @@ export default function Main() {
                     ))}
                 </List>
             </Drawer>
-            <main className={classes.content}>
+            <main className={classes.content} style={{ backgroundColor: 'red' }}>
                 <div className={classes.toolbar} />
-                <Typography paragraph>
-                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
-                    incididunt ut labore et dolore magna aliqua. Rhoncus dolor purus non enim
-                    praesent elementum facilisis leo vel. Risus at ultrices mi tempus imperdiet.
-                    Semper risus in hendrerit gravida rutrum quisque non tellus. Convallis convallis
-                    tellus id interdum velit laoreet id donec ultrices. Odio morbi quis commodo odio
-                    aenean sed adipiscing. Amet nisl suscipit adipiscing bibendum est ultricies
-                    integer quis. Cursus euismod quis viverra nibh cras. Metus vulputate eu
-                    scelerisque felis imperdiet proin fermentum leo. Mauris commodo quis imperdiet
-                    massa tincidunt. Cras tincidunt lobortis feugiat vivamus at augue. At augue eget
-                    arcu dictum varius duis at consectetur lorem. Velit sed ullamcorper morbi
-                    tincidunt. Lorem donec massa sapien faucibus et molestie ac.
-                </Typography>
+                <Card />
                 <Typography paragraph>
                     Consequat mauris nunc congue nisi vitae suscipit. Fringilla est ullamcorper eget
                     nulla facilisi etiam dignissim diam. Pulvinar elementum integer enim neque
